Support for laravel 7

+++ database/migrations/2016_07_29_171118_create_chatter_discussion_table.php
@@ -9,9 +9,9 @@ class CreateChatterDiscussionTable extends Migration
     {
         Schema::create('chatter_discussion', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('chatter_category_id')->unsigned()->default('1');
+            $table->bigInteger('chatter_category_id')->unsigned()->default('1');
             $table->string('title');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->boolean('sticky')->default(false);
             $table->integer('views')->unsigned()->default('0');
             $table->boolean('answered')->default(0);
